Feat: Handle MultisigRejectionChanged
