+++ src/handlers/multiSig/index.ts
@@ -2,5 +2,6 @@ export * from './multiSigCreated/multiSigCreated';
 export * from './globalThresholdSet';
 export * from './domainSkillThresholdSet';
 export * from './approvalChanged/approvalChanged';
+export * from './rejectionChanged/rejectionChanged';
 export * from './motionCancelled';
 export * from './multiSigExecuted/multiSigExecuted';
